requirements.txt
tensorflow
numpy
Flask
Pillow
logging
cassandra-driver

